# pie-slice/openapi.yaml
openapi: 3.1.0
info:
  title: pie-slice API
  version: 0.2.0
  description: |
    Backend the `frontend/` React app expects, derived directly from
    `frontend/src/services/types.ts` (the `ExpenseServiceApi` interface —
    the single boundary every UI call crosses) and
    `frontend/src/services/mockService.ts` (the no-backend reference
    implementation this spec must behave identically to).

    ## Authentication

    Real accounts: email + password, bearer tokens. `POST /auth/signup`
    and `POST /auth/login` are the only unauthenticated endpoints; every
    other endpoint requires `Authorization: Bearer <token>`, obtained from
    either of those two.

    Beyond "is this a valid token," authorization is enforced per
    product-spec.md:
    - **Group access**: only a *linked* member of a group may view or act
      on it. A request for a group you're not a linked member of gets the
      same `404` as a group that doesn't exist — existence is never
      leaked to non-members.
    - **Edit/delete on an expense or settlement**: only the user who
      created it (`createdByUserId`) may edit or delete it — enforced
      with `403`, checked server-side on every request, never trusted
      from the client.
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /groups/{groupId}/settlements:
    get:
      tags: [settlements]
      operationId: listSettlements
      summary: List a group's settlements
      description: Maps to `ExpenseServiceApi.listSettlements`.
      parameters:
        - $ref: "#/components/parameters/GroupId"
      responses:
        "200":
          description: All settlements for the group.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Settlement"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [settlements]
      operationId: addSettlement
      summary: Record a settlement
      description: >
        Maps to `ExpenseServiceApi.addSettlement`. Records that
        `fromMemberId` paid `toMemberId`; balances shift accordingly (see
        `/groups/{groupId}/balances`). Not gated on the settle-up
        suggestion — any from/to/amount combination is accepted.
        `createdByUserId` on the response is always the caller.
      parameters:
        - $ref: "#/components/parameters/GroupId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettlementInput"
      responses:
        "201":
          description: Settlement recorded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settlement"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /groups/{groupId}/settlements/{settlementId}:
    put:
      tags: [settlements]
      operationId: updateSettlement
      summary: Update a settlement
      description: >
        Maps to `ExpenseServiceApi.updateSettlement`. Full replacement,
        matches `UpdateSettlementInput = AddSettlementInput`.
        **Creator-only** — see Authentication above.
      parameters:
        - $ref: "#/components/parameters/GroupId"
        - $ref: "#/components/parameters/SettlementId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettlementInput"
      responses:
        "200":
          description: Settlement updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settlement"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/NotCreator"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [settlements]
      operationId: deleteSettlement
      summary: Delete a settlement
      description: Maps to `ExpenseServiceApi.deleteSettlement`. **Creator-only** — see Authentication above.
      parameters:
        - $ref: "#/components/parameters/GroupId"
        - $ref: "#/components/parameters/SettlementId"
      responses:
        "204":
          description: Settlement deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/NotCreator"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Token returned by `POST /auth/signup` or `POST /auth/login`.
  parameters:
    GroupId:
      name: groupId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SettlementId:
      name: settlementId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Missing, invalid, or expired bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: >
        The group (or the expense/settlement within it) does not exist,
        or you're not a linked member of the group. Mirrors the mock's
        `getGroup` returning `null` / other methods throwing "not found"
        errors — see Authentication above for why membership and
        existence collapse to the same response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotCreator:
      description: You are not the user who created this expense/settlement, so you may not edit or delete it.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: >
        Request body failed schema validation (FastAPI's standard
        automatic pydantic validation error format).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPValidationError"
...
  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          example: "Percentages sum to 90%, but must sum to 100%."
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
              msg:
                type: string
              type:
                type: string
...
    SettlementInput:
      description: Request body shared by create (POST) and full-replacement update (PUT). Matches AddSettlementInput / UpdateSettlementInput in services/types.ts.
      type: object
      required: [fromMemberId, toMemberId, amountCents, date]
      properties:
        fromMemberId:
          type: string
          format: uuid
        toMemberId:
          type: string
          format: uuid
        amountCents:
          type: integer
          exclusiveMinimum: 0
        date:
          type: string
          format: date
        method:
          type: string
          description: Optional free-text payment method, e.g. "Venmo", "cash".
    Settlement:
      type: object
      required: [id, groupId, fromMemberId, toMemberId, amountCents, date, createdByUserId]
      properties:
        id:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
        fromMemberId:
          type: string
          format: uuid
        toMemberId:
          type: string
          format: uuid
        amountCents:
          type: integer
        date:
          type: string
          format: date
        method:
          type: string
        createdByUserId:
          type: string
          format: uuid
          description: Whoever was logged in when this was created — only they may edit/delete it.
...